Adde Challenge-Participant Registration Questions (#3528)
This PR is linked to the pitch:
- https://github.com/DIAGNijmegen/rse-roadmap/issues/344

It hooks in showing and getting answers to the questions from:
- https://github.com/comic/grand-challenge.org/pull/3526

Note that `""` is not technically a JSON (`""""` is the empty string in
JSON notation). However, having users input everything with double
quotes seems very, very silly. As such the `JSONField` is wrapped to
correctly parse strings.