Fix cleaning of documentation pages (#3567)
The content of a documentation page is markdown. Cleaning the markdown
will escape characters such as `>`, which means they then fail to render
correctly in code blocks. This step is anyway unnecessary: we save the
raw markdown and then convert it on the fly in the template which calls
`md2html`. At no point do we assume that `content` is safe, so it is
okay to remove this step.