Refactor job creation code (#3465)
This PR refactors the job creation code. Jobs, archive items and display
sets now all use the same creation and validation logic.

Errors from asynchronous input validation now get caught and are
reported back to the user as error message on the Job object. The job is
then cancelled and not executed.

This PR also enables selecting existing images and files for jobs
through the UI and enables re-using CIVs for existing values if
possible. The same holds partially for the API view as well. It is not
yet possible to create a job through the API with an existing non-image
file. This was out of scope for this PR
(https://github.com/DIAGNijmegen/rse-roadmap/issues/335#issuecomment-2304484574)

Part of https://github.com/DIAGNijmegen/rse-roadmap/issues/335 

Closes https://github.com/comic/grand-challenge.org/issues/3139
Closes https://github.com/comic/grand-challenge.org/issues/3325
Closes https://github.com/comic/grand-challenge.org/issues/3565
Partially addresses
https://github.com/comic/grand-challenge.org/issues/3368

---------

Co-authored-by: James Meakin <12661555+jmsmkn@users.noreply.github.com>