Consolidate component job requirement fields (#3616)
This PR consolidates `requires_gpu` and `desired_gpu_type` into one
choice field `requires_gpu_type` and sets this on the algorithm and
phase rather than on the individual container images along with
`requires_memory_gb`. This is because they rarely change between
container images for the same algorithm and users more often forget to
update.

I removed the option for challenge admins to set a gpu for an evaluation
as this has not been used in 3 years. We can still do that from the
admin.

See https://github.com/DIAGNijmegen/rse-roadmap/issues/346