Fix auto accepting pending requests upon verification (#3659)
This fix relates to issue #2461 and PR #3514. The issue is that the
logic of auto accepting pending requests upon verification was
implemented in the save() method of the model. However, in Admin, the
actions are performed as an update on the queryset which does not
trigger the save() method, and similar behaviour affected the proper
handling of permissions since the queryset update did not trigger the
pre_save signals.