Prevent duplicate notifications for failed image imports (#3672)
I noticed while doing some more local testing for the workshop that
failed image uploads (uploading an image that none of our importers can
read) results in 2 notifications being sent to users, one from
`build_images` and one in `add_image_to_object`.

This PR adds a check on the upload session status at the beginning of
`add_image_to_object`, making sure that we exit this task if the upload
session failed. We also check in this task that there is an image for
the upload session and we could in theory exit the task there (rather
than sending a notification, like we do now), but that seems too
indirect of a check to me, since there might be other reasons that the
image does not exist (anymore), independent of the upload session
status.

While working on this, I noticed that we mark uploads as successful even
if 0 images could be imported. That doesn't really make sense. I checked
in the codebase for dependencies on that assumption, but didn't find
any, so I'm proposing to mark those as failed.