Update challenge cost calculation (#3696)
This updates the cost calculation in line with our new policy and
removes the cost calculation from the request form.
The cost estimate in the detail view has been updated to include the
warnings and extra calculations that were previously only displayed in
the request form.

Part of https://github.com/DIAGNijmegen/rse-roadmap/issues/356

The updated cost estimate on the detail view (only visible to reviewers)
now looks as follows:


![image](https://github.com/user-attachments/assets/b258bb80-8180-41a0-82b9-f2f0620f5814)