Adds adjustable gpu types for algorithms (#3714)
Make the selectable gpu type and maximum memory setting for algorithms
adjustable using Phase settings. Relevant phases are filtered when
updating or creating an algorithm, this is based on inputs and outputs
match, participant match, the phase submissions kind must be algorithm,
and the phase must be public.

The new Phase settings are also used to validate algorithm gpu and
memory settings when it is submitted to the phase.

See https://github.com/DIAGNijmegen/rse-roadmap/issues/346