Update `JSONEditorWidget` to 10.1.1 (#3728)
This PR has a few changes bundled with respect to the JS-dependency
JSONEditorWidget.

## Amass

- Update `amass.lock`, this was outdated with respect to the
`pyproject.toml`, a lot of no-longer-defined dependencies are present
- Update JSONEditor to the latest version (`10.1.1`)

## JsonEditor invocation

Fixes #3698

The crux of the original issue lies with `setCustomValidty` on a
`d-none` classed `<textarea>`. When saving, it has the client-side
validity crash since it cannot focus on the element.

To solve this the validity error is now set on the textarea of the ace
editor:


![image](https://github.com/user-attachments/assets/5fd1a3ed-62b9-43b0-94ec-f0fd35f9d933)


Another (hidden) error would occur when trying to switch from an empty
"code" mode to "tree" mode. With the latest version of JSONEditor this
now shows as an informative (albeit ugly) prompt:

![image](https://github.com/user-attachments/assets/88b3ee11-0f8e-4b8c-8aef-9907ebdda325)

---------

Co-authored-by: James Meakin <12661555+jmsmkn@users.noreply.github.com>