Retain current selection of gpu type choice and allow challenge admins to update gpu types (#3734)
Retain the current selection of `job_requires_gpu_type` as an option for
choices of this field on the Algorithm update form. Otherwise, when the
options are changed (in the relevant phases or organization or because
the user has changed) the current selection can easily be lost if it's
not included as an option on the field. (The form will then
automatically select the first available option without informing the
user.)

We no longer filter only on public phases for the options of selectable
gpu type choices. This is to allow challenge admins to test their
algorithms while the challenge is still hidden.

Related to https://github.com/DIAGNijmegen/rse-roadmap/issues/346