Replace responsive DataTables behaviour with floating scrolling bar (#3729)
Fixes #3028

This PR adds a new JS dependency, `floating-scroll`. This tiny JS
package allows a floating scrollbar at the bottom of a viewport instead
of the bottom of the element (which is possibly hidden way below).

Floating-scroll is applied after the Datatables are initiated. This is
done via a generic listener to `init.dt` on the document level: in the
same breath, it sets the `scrollX: true` to allow the inner element to
overflow and spawn a scrollbar in the first place.

Datatables come in two flavours:
- Dynamically, initialized via the partial template `datatable.html`
(and some Python, JS et cetera)
- Statically, by initializing it on a `<table>` directly

The new overflow behaviour applies to both. _For the static route_ this
entails the scrollbar is properly under the table now instead of under
the paging UI. The same holds for the _dynamic route_ but this also has
the `responsive` component explicitly removed.

## Visuals
### Current
 

https://github.com/user-attachments/assets/1479346b-69a2-4e51-903a-73e48a87ea7c


https://github.com/user-attachments/assets/393d12f0-1513-4800-977e-e70802eaddfb

### New


https://github.com/user-attachments/assets/189a9cdd-f42d-4d39-9328-2fb94999d465


https://github.com/user-attachments/assets/393a8f48-68d8-4e28-bdc8-581917f0c3c4


### Side note

Technically, the bootstrap class `"table-responsive"` is now redundant:
the element containing the datatable should never overflow—unless
perhaps there are a gazillion pages(?). However, since it's generally
applied as a wrapping element, removing it here would result in an
unneedingly bloated changeset.

I first had handy-scroll, but that has a component-only version 2 now...
which does not work with dynamic datatables.

---------

Co-authored-by: James Meakin <12661555+jmsmkn@users.noreply.github.com>