Upgrade Datatables (#3738)
The last update to Datatables was from a release that was from 24th
September, 2021. Since I had my hands in the Datatables last week I
thought now would be a good time to bump the version.

The latest (2.1.8) has jquery as a hard dependency, so no need to
explicitly select that any more.

It was nearly a perfect drop-in replacement, with the exception of:
- The `ajax.url` does not default to the current HREF when set to `""`.
Rather `"."` does.
- Sorting now has three stages (was two: asc/desc). The addition is a
neutral sorting, which excludes the `order[...]` XML arguments in the
xhr request. Instead, I've made it so that the data ordering falls back
to the default found on the implementing class. Imho that is superior to
random ordering.