Fix several issues with Datatables (#3750)
This addresses several issues for the styling of Datatables

Closes #3745

## Styling

The PR introduces a few custom CSS styles, overwriting some of the
styles found in the bootstrap-4 extension of Datatables.
 
## Leaderboard
The `dom` option is deprecated in Datatables. It had to be updated
anyway to account for the new styling. This PR uses the new `layout`
option in Datatables to set things up in a future-proof manner.

There are four states, as shown below.

#### Regular

![image](https://github.com/user-attachments/assets/7657b044-1b5d-4f87-97bc-042a3fc77fc9)

#### Date & Additional Metrics

![image](https://github.com/user-attachments/assets/ce2f7ab3-b8e4-4823-a864-cfaf0fca4f4b)

#### Date

![image](https://github.com/user-attachments/assets/65494f38-6651-40fa-85dc-d518a35da5c0)

#### Additional Metrics

![image](https://github.com/user-attachments/assets/67774a01-80a9-4863-8b35-aa710c8332a4)


### Other
It also consolidates a few common settings in Datatables:
- Adds a default via-html-element-class selectable `"nonSortable"`
column, employing it in one additional location.
- Sets the default paging size to 50 (most common).
- Sets the default type to `"string"` of client-side datatables, which
indicates: do not transform content when searching/ordering. Also
prevents unexpected styles from being applied because of 'automagically'
detected content types.