Tweak image upload help text (#3774)
A user mentioned some confusion by clinicians trying to upload dicoms.
Imho the user had a valid point. The help text also had the suggestion
that .tiff was converted to MHA. That is incorrect. I opted to sacrifice
some horizontal space for clarity.

## Before


![image](https://github.com/user-attachments/assets/8dfde665-f2da-493f-8bd4-ff39b6fde567)

## After


![image](https://github.com/user-attachments/assets/3c2ef64a-d1ff-44a5-bf84-5db5ca8057e2)

Note: text is used in a widget, no need for a migration.