Unmix presentation and content for docpages (#3786)
This adds css for docpages, to enable writing simple markdown focused on
content rather then presentation.

For example this removes the need to write 
`<h3 class="text-center">heading text</h3>`
to create a centered heading at the top of the page. 
This can now be written simply as 
`### heading text`