Move input/output annotation from manager to separate method (#3804)
I added an input/output count annotation to the
`AlgorithmInterfaceManager` in my last PR, but had not tested the
migrations again. In my migration check for the phase interfaces, I
realized that the migrations break with that addition since they don't
have access to custom manager methods. So I'm moving the annotation to a
separate function here. The upside of this is that I can now use it in 2
other places where we're making the same annotation. So I'm calling this
a win.

Note that this does not add or change functionality, it just moves the
code. So I didn't spend time on adding a test for the new function - it
is indirectly covered by tests of all the places where it is used.