Rename interface to socket on user-facing views (#3846)
This renames `interface` to `socket` on user-facing views and forms. 
I did not change the urls that include the word `interface`. That seems
less important to me for now. We will need to add permanent redirects
when we do change them.

I based this on the feature branch for optional inputs since a lot of
the places that require changes are touched in this branch. The users
can wait 1 week for the renaming I think.

The only places where the user is still confronted with the word
`interface` now (as far as I can tell) is the API. Updating that is more
complex.