Split off the Ground Truth page (#3853)
Part of the pitch:
- https://github.com/DIAGNijmegen/rse-roadmap/issues/385

A quick pre-lude:
- split the ground truth from being a tab in the the reader study detail
to its own view
- add a warning when the reader study is not configured for educational
purposes

Other than above, I did not change the content of the page.

Considered to refactor all reader study views to have the configuration
side-panel, much like the algorithms. But have put that out of scope
(for now).



## Showcase

The warning is no longer there.


![afbeelding](https://github.com/user-attachments/assets/f88398bc-e5b5-4f7f-99be-6edc538ac9ae)