Cleanup flexible widget (#3833)
Cleanup after changes for the flexible file widget to make future
maintenance easier.

`InterfaceFormField` has been renamed to `InterfaceFormFieldFactory` and
has been stripped of much of the logic, which had become very confusing.

User object permissions are now checked before rendering, to avoid
leaking information about object existance.

This also contains a functional change to the htmx for the
`FileSearchWidget` and the `ImageSearchWidget`.
`hx-trigger` now includes `input` ensure triggering when text is
removed.
`hx-swap` is now the innerHTML of the select element. This makes it
easier to see that this element has the widgetname which is where the
value is taken from.

Related to https://github.com/DIAGNijmegen/rse-roadmap/issues/363

---------

Co-authored-by: James Meakin <12661555+jmsmkn@users.noreply.github.com>