Add postpaid invoices (#3855)
Add post-paid invoices. Compute costs on postpaid invoices are
immediately available, regardless of the payment status, but only if
there is another non-complimentary invoice in the paid state.

- add payment_type field (pre-paid, post-paid, or complimentary)
- This moves COMPLIMENTARY from invoice PaymentStatusChoices to
PaymentTypeChoices.
  - Involves data migration
- add new payment status "Invoice requested"
- update approved_compute_costs_euro_millicents to include post-paid
invoices if there is a non-complimentary invoice in PAID state

This also adds text for requesting the invoice in the django admin, auto
populated from the field contents.

Related to https://github.com/DIAGNijmegen/rse-roadmap/issues/373

---------

Co-authored-by: James Meakin <12661555+jmsmkn@users.noreply.github.com>