Clear existing phase inputs (#3865)
All phases have default input interfaces defined at the moment, but the
current workflow ignores them. We now want to make it possible to add
inputs that should not be ignored. For that, it makes sense to remove
those default inputs and to only set inputs in the future when those
should actually be shown as additional inputs for users when creating a
submission.

Part of https://github.com/DIAGNijmegen/rse-roadmap/issues/357