Seperate evaluation creation from execution (#3866)
This separates the creation of the evaluation from its execution (and
for algorithm submission, from scheduling the algorithm jobs). This is
necessary because we will want to validate and attach additional inputs
to the evaluation object before this happens. The validation and
attaching of these additional inputs will be added in another PR.

Part of https://github.com/DIAGNijmegen/rse-roadmap/issues/357