Ground Truth csv form seperation (#3879)
Part of the pitch:
- https://github.com/DIAGNijmegen/rse-roadmap/issues/385

This PR further pull apart the ground truth main view and the csv upload
form.

This is to create a nice niche where the answer copy form can be placed
in both UI and workflow.

It also:
- Fixes having non ground-truth applicable questions in the example csv
(e.g. annotations)
- Adds a warning for 'incomplete' ground truth that should give some
warnings to reader study admins when they added a new question or have
forgotten to set the ground truth when it is set to educational.
- Add a confirmation modal when deleting GT

## Showcase



https://github.com/user-attachments/assets/8ff85dc5-e6f6-4d02-9d7b-1c099394afbc