Ground truth copy from answers (#3884)
Part of the pitch:
- https://github.com/DIAGNijmegen/rse-roadmap/issues/385

This PR adds a form that, when submitted, copies the answers of a
reader/editor as ground truth.

Only users that have provided answers are selectable. However, a 100%
completion of the reader study is required to consider the select user
valid.

Since readers are limited in number I've explicitly opted to not use an
autocomplete view + widget and have instead used a simple Select.


# Showcase


https://github.com/user-attachments/assets/2c1d21d0-8177-4aee-9ffe-308938fcf0ac

---------

Co-authored-by: James Meakin <12661555+jmsmkn@users.noreply.github.com>