Image processing and assignment: account for object not existing (#3930)
Closes #3929

There are three steps involved in image processing:

1. `RawImageUploadSession.process_images` (sync)
    A. Interval
2.  `tasks.build_images` (async)
    B. Interval
3. `task.assign_image_to_object` (async)

During both interval A and B the target `DisplaySet` / `ArchiveItem`
could be removed by the user.

This PR adds some handling for that: it should gracefully ignore it when
it refers to `DisplaySets` / `ArchiveItems`, never a `Job`.

It also applies the same logic to the adding of files.

---------

Co-authored-by: James Meakin <12661555+jmsmkn@users.noreply.github.com>