Ground Truth from Answers (#3922)
This PR is the final part on:
- https://github.com/DIAGNijmegen/rse-roadmap/issues/385

It adds a form that allows the ground truth to be transformed into
answers for a selected user. The selected user should not have any
existing answers.

To simplify the permissions I've opted to fully create new answers in an
asyncronous task. However, because of wish to have the whole conversion
in a single transaction I ran into problems with signals.

As a consequence I've also opted to boot the signal on the post_save of
an `Answer` and pull the logic there into the `Answer.save()`. It used
to be the case that an async task got spawned for each new `Answer`
added. That is now done synchronously: the ground truth was being
retrieved anyway. When added ground truth, which is done in bulk anyway,
the re-calculation is now batched in a single task. Should be a bit more
performant than a typical 1000 tasks spawned instantly.

Not sure about the current flag-based calculation of scores approach to
`Answer.save()`: perhaps there is a better ay? It fixes an N+1 run off
on creating answers from ground truth.

## Showcase

Showing the loop: answers to ground truth to answers to ground truth





https://github.com/user-attachments/assets/bd4f2915-6eb5-4a79-9a0b-ac888b3d519a

---------

Co-authored-by: James Meakin <12661555+jmsmkn@users.noreply.github.com>