Fix celery exception handling with direct invocation (#3992)
We have the option `ignore_errors=(...,)` for celery tasks that is set
so that the task gets marked as successful if one of those errors is
raised. However, when invoking the task directly those errors should
still be raised so that the caller can then process the error. This can
be in a try-except block, or by adding it to it's own `ignore_errors`.

This prevents the following:

```python
@task(ignore_errors=(Foo,))
def a():
    raise Foo

@task(ignore_errors=(Foo,))
def b():
    a()
    do_something_assuming_a_ran_successfully()
```

This PR will ensure that if `b` is invoked that
`do_something_assuming_a_ran_successfully` will not run, but the task
will still be marked as successful as the raised `Foo` is ignored.