Only use union in permission filter when necessary (#4026)
This PR allows using the most optimal filter in all cases. By defining
which permissions are allowed on the user and group models we can avoid
having to search over both tables. Now we should be able to see the
cases where this is necessary - currently these are "change_job",
"view_image", "view_challengerequest", "view_submission", and "view_answer".

The allowed_permissions have been populated by inspecting the production
database. If they are any extra ones then this only affects performance.
Errors will be raised if permissions are attempted to be assigned that
are not allowed.