Allow model-based POSTs to create SocketValues (#4030)
In rse-gcapi we have this definition of a request:
```Python
@dataclass
class ComponentInterfaceValuePostRequest(BaseModel):
    interface: str
    value: Optional[Any]
    file: Optional[bytes]
    image: Optional[str]
    upload_session: Optional[str]
    user_upload: Optional[str]
 ```
 
However, if one would try to submit using this model as a base we run into trouble with fields which are not allowed to be null.
 
This PR cleans up the fields and validation of the PostRequest as to account for some of the fields to be there but have a null value.
 
 This opens up the option to use local-validated models when POSTing from gcapi:
 
 ```Python
 
  c.archive_items.partial_update(
        pk=ai.pk,
        values=[
            ComponentInterfaceValuePostRequest(
                interface=socket.slug,
                value="foo",
                file=None,
                image=None,
                upload_session=None,
                user_upload=None,
            )
        ],
    )
```