Shorten paths found in starter-kit / algorithm templates (#4059)
Related:
-https://github.com/DIAGNijmegen/rse-grand-challenge-forge/issues/92

This PR updates to latest forge version that has some shorted path
names.

It also hides the options to download the kit if there are no interfaces
configured.

Finally, it removes the 'starter' from 'starter-kit' in the generated
.ZIP file. It had already removed adding the containing directory:
however several unzipping apps tend to automatically add the name of the
filename as holding directory if there is not a singular folder/item in
the ZIP. Which makes sense but this does add to the path length.

I think that was all the 'fat' that was on the path lengths in the forge
and it can't be made much leaner. When the forge gets pulled into GC we
should likely rewrite the generation to work on the Phase level. It
currently operates on the Challenge level with a singular phase in the
context. That would allow us to also remove the intermediate phase
folder.

I'll mention unpacking is most wise under Linux env in the doc page on
the starter kit.