Use window annotation for standard index (#4098)
Follow on from #4097, this saves having to lookup the index of a display
set by having them all in memory. This only works when **not** shuffling
the hanging list, that is much more difficult to change.

For shuffled indexing an annotation is now added to each object, rather
than having to perform an array search for each object when the index is
required.

I did think of creating the shuffled index in python and then looking up
by the row number, however, we cannot use the row number in another
annotation. Another strategy would be needed to do that (perhaps saving
the index on the display set, however, then you need to update all other
display sets for the reader study on deletion). Likewise, answerable
questions could also be moved to the database.

This is already covered by tests that kept breaking, so fairly confident
in the implementation.