Replace datatable initialization and configuration javascript with data attributes (#4193)
While working on the javascript conversion I notices lot's of simple
javascript files that contained only a few lines for initialization of
DataTable. In the new setup these would all require separate endpoints
to be built by npm, but that is quite a big overhaul for such a trivial
feature. Therefore, I decided to create a helper function that
initializes and configures DataTable when certain data attributes are
present on the table element. This allows for easy configuration without
needing to write JavaScript. To enable this add the `data-data-table`
attribute on the table element. Then other attributes prefixed with
`data-dt-` can be used to configure DataTable.

 Example:
 ```html
<table data-data-table data-dt-page-length="50" data-dt-order='[[0,
"asc"]]'>
 ```