Refactor object locking (#4303)
This PR refactors how we perform object locking. It introduces a context
manager, `check_lock_acquired`, that will transform the operational
error to a lock not acquired exception for use internally. This saves
having to do the large construction.

The PR also removes `lock_model_instance`. This method turned out to be
quite complex with all the locking options, so now we can use the Django
constructs more closely.

The method also encouraged us to use `get_model`, which isn't a great
practise as it uses string imports. Now we can use explicit imports
everywhere possible which should improve the experience when working in
IDEs.

Finally, I noticed that `object` was being shadowed in too many places
so corrected that.