Alter DICOM endpoint response (#4334)
Return an array of objects for the image frames which allows for more
flexibility in the future. Adds the image set and image frame ids for
cross-referencing.

Adds the accept and accept encoding headers to ensure the responses are
consistent.

Stores extra metadata that is required for fetching of each image frame.

See https://github.com/DIAGNijmegen/rse-roadmap/issues/432