Refactor ValuesForInterfacesMixin to LinkedComponentInterfaceMixin (#4354)
Remove values from `ValuesForInterfacesMixin` because we only use the
keys in `values_for_interfaces`, i.e. the component interfaces, and
return the interfaces as a queryset instead of converting them to a
list.

Use `linked_component_interfaces`, which returns a queryset. 
Change `linked_component_interfaces` on `Algorithm` and `Phase` to
return a queryset (was a set) as well, to avoid confusion over return
type for properties with the same name.