Avoid server error in FlexibleImageWidget validation (#4381)
Fix error in `FlexibleImageWidget.decompress()` due to unexpected value.
This now properly raises a "field required" validation error on the
field.

The possible data returned from the possible widgets of
`FlexibleImageWidget` was not completely tested. This adds a test where
we extract the form data from the response of the flexible image widget
for each of the possible options the user can select.

Closes https://github.com/comic/grand-challenge.org/issues/4312