Update image permission assignment through algorithm jobs (#4386)
Avoids the use of the slow `update_viewer_groups_permissions` method by
assigning the `view_image` permissions directly. The downside of this is
that the permissions will need to be kept in sync in both `signals.py`
and `_get_expected_job_viewer_groups` but we have good test coverage for
this.

Closes
https://github.com/DIAGNijmegen/rse-grand-challenge-admin/issues/663