Quick fix for errors on running celery tasks without celeryd (#4401)
The uv-based integration tests for gcapi run without workers. This
entails that the `task_prerun` and `task_postrun` listeners fire, but
not the `celeryd_after_setup`.

As a result, this throws an error several times during setup (but
isolated so nothing breaks):
> AttributeError: 'Celery' object has no attribute 'is_solo_worker'

Instead of re-adding the worker, I thought it's OK to instead have the
prerun/postrun to take into account the attr hadn't been set.